requirements.txt
Flask
pandas
pyodbc
python-dotenv
gunicorn
redis

